backend/requirements.txt
Django==2.2.19
pytz==2023.3
sqlparse==0.4.4
asgiref==3.2.10
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest
pytest-django
pytest-pythonpath
djoser==2.1.0
django-colorfield==0.7.2
drf-extra-fields==3.4.0
drf-yasg==1.21.3
django-rest-swagger==2.2.0
python-dotenv==0.21.0

